6.0 /BDSM > Removed /Manifest > Added
Manifest is an automapper with 3 options. Hunt for exp, mine for ore which awards coins, or gather for materials.

your echelon and equipped tarot star level are the two main factors affecting the reward outcomes

The cooldown is 6 hours + your echelon. subject to change. I may add further bonuses to certain cards or have the card tier affect the cooldown time etc.